defining eplus base model
