+++ opyplus/weather_data/design_day.py
@@ -2,7 +2,7 @@
 
 from ..epm.parse_idf import parse_idf
 from ..epm.table import Table
-from ..epm.epm import Epm
+from ..epm.epsf import Epsf
 import collections
 from ..epm.relations_manager import RelationsManager
 from .. import CONF
@@ -10,12 +10,14 @@ import logging
 
 logger = logging.getLogger(__name__)
 
-DDY_TABLE_DESCRIPTORS_REF = ["sizingperiod_designday",
-                             "site_location",
-                             "runperiodcontrol_daylightsavingtime"]
+DDY_TABLE_DESCRIPTORS_REF = (
+    "sizingperiod_designday",
+    "site_location",
+    "runperiodcontrol_daylightsavingtime",
+     )
 
 
-class Ddy(Epm):
+class Ddy(Epsf):
     """
     Ddy model.
 
@@ -41,30 +43,19 @@ class Ddy(Epm):
 
     Notes
     -----
-    Ddy files are not versioned, and by default will use latest IDD version for model conversion from .ddy
+    Ddy version behaviour
+        Ddy files are not versioned, and by default will use latest IDD version for model conversion from .ddy
     """
 
-    def __init__(self, json_data=None, check_required=True, check_length=True, idd_or_version=None):
+    def __init__(self, json_data=None, check_required=True, check_length=True, idd_or_version=None,table_refs_selection=DDY_TABLE_DESCRIPTORS_REF):
         # call super
         super().__init__(
             json_data=json_data,
             check_required=check_required,
             check_length=check_length,
-            idd_or_version=idd_or_version
+            idd_or_version=idd_or_version,
+            table_refs_selection=table_refs_selection
         )
-        # !! relations manager must be defined before table creation because table creation will trigger
-        # hook registering
-        self._dev_relations_manager = RelationsManager(self)
-
-        self._tables = collections.OrderedDict(sorted([  # {lower_ref: table, ...}
-            (table_descriptor.table_ref.lower(), Table(table_descriptor, self))
-            for table_descriptor in self._dev_idd.table_descriptors.values()
-            if table_descriptor.table_ref.lower() in DDY_TABLE_DESCRIPTORS_REF
-        ]))
-
-        # load json_data if relevant
-        if json_data is not None:
-            self._dev_populate_from_json_data(json_data)
 
     # --------------------------------------------- public api ---------------------------------------------------------
     # python magic
@@ -74,13 +65,13 @@ class Ddy(Epm):
 
         Returns
         -------
-        {'<Epm>'}
+        {'<Ddy>'}
         """
         return "<Ddy>"
 
     def __str__(self):
         """
-        Str representation of Epm, with the number of records per tables.
+        Str representation of Ddy, with the number of records per tables.
 
         Returns
         -------
@@ -103,16 +94,19 @@ class Ddy(Epm):
 
         Returns
         -------
-        dict
+        design_day_dict
         """
-        # create data
+        design_day_select = self.sizingperiod_designday.select(lambda x: design_day_ref in x.name)
+        if len(design_day_select) == 0:
+            raise ValueError(f"Design day with name '{design_day_ref}' not found.")
+        if len(design_day_select)>1:
+            raise ValueError(f"Ambiguous design day selection, {len(design_day_select)} design days with name '{design_day_ref}' were found, be more concise.")
+        
         design_day = self.sizingperiod_designday.one(lambda x: design_day_ref in x.name)
-        design_day_dict = {design_day.get_field_descriptor(field).ref: design_day[field] for field in
-                           range(len(design_day))}
-
+        design_day_dict = {design_day.get_field_descriptor(field).ref: design_day[field] for field in range(len(design_day))}
         return design_day_dict
 
-    def copy_to_epm(self, epm, design_day_ref):
+    def add_design_day_to_epm(self, epm, design_day_ref):
         """
         Add design day to Epm
         """
@@ -139,5 +133,6 @@ class Ddy(Epm):
         return cls._create_from_buffer_or_path(
             parse_idf,
             buffer_or_path,
-            idd_or_version=CONF.default_idd_version  # .ddy are not version: latest idd is used
+            idd_or_version=CONF.default_idd_version,  # .ddy are not version: latest idd is used
+            table_refs_selection=DDY_TABLE_DESCRIPTORS_REF
         )
